# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Pourier API'
  description: "API REST pour la plateforme de vente de photos Pourier. Cette API permet aux utilisateurs d'acheter des photos, aux photographes de vendre leurs créations, et aux administrateurs de gérer la plateforme."
  contact:
    name: 'Pourier Support'
    email: contact@pourier.com
  license:
    name: Proprietary
    url: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'Serveur API Pourier'
paths:
...
  /api/orders:
    get:
      tags:
        - Orders
      summary: "List user's orders"
      description: 'Get all orders for the authenticated user with pagination'
      operationId: getOrders
      parameters:
        -
          name: per_page
          in: query
          description: 'Number of orders per page'
          required: false
          schema:
            type: integer
            default: 20
            example: 20
      responses:
        200:
          description: 'Orders retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                  links:
                    $ref: '#/components/schemas/PaginationLinks'
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Orders
      summary: 'Create new order'
      description: 'Create a new order with items from cart. Platform takes 20% commission.'
      operationId: storeOrders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - items
                - subtotal
                - total
                - payment_method
                - billing_email
                - billing_first_name
                - billing_last_name
                - billing_phone
              properties:
                items:
                  type: array
                  items:
                    properties:
                      photo_id:
                        type: string
                        format: uuid
                        example: 9d445a1c-85c5-4b6d-9c38-99a4915d6dac
                      license_type:
                        type: string
                        enum:
                          - standard
                          - extended
                        example: standard
                    type: object
                  example:
                    -
                      photo_id: 9d445a1c-85c5-4b6d-9c38-99a4915d6dac
                      license_type: standard
                subtotal:
                  description: 'Sum of all item prices in FCFA'
                  type: number
                  format: float
                  example: 15000
                tax:
                  description: 'Tax amount (currently 0)'
                  type: number
                  format: float
                  example: 0
                discount:
                  description: 'Discount amount'
                  type: number
                  format: float
                  example: 0
                total:
                  description: 'Total amount in FCFA'
                  type: number
                  format: float
                  example: 15000
                payment_method:
                  type: string
                  enum:
                    - mobile_money
                    - card
                  example: mobile_money
                billing_email:
                  type: string
                  format: email
                  example: user@example.com
                billing_first_name:
                  type: string
                  example: Jean
                billing_last_name:
                  type: string
                  example: Dupont
                billing_phone:
                  description: 'Burkinabé phone format'
                  type: string
                  example: '+226 70 12 34 56'
              type: object
      responses:
        201:
          description: 'Order created successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Commande créée avec succès'
                  data:
                    $ref: '#/components/schemas/Order'
                type: object
        400:
          description: 'Order creation failed - photo unavailable or validation error'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Erreur lors de la création de la commande: La photo 'xyz' n'est plus disponible"
                type: object
        401:
          $ref: '#/components/responses/UnauthorizedResponse'
        422:
          $ref: '#/components/responses/ValidationErrorResponse'
      security:
        -
          bearerAuth: []
...
  /api/webhooks/cinetpay:
    post:
      tags:
        - Webhooks
      summary: 'CinetPay payment webhook (callback)'
      description: 'Receives payment notifications from CinetPay when payment status changes. Uses HMAC signature verification for security. This endpoint is called by CinetPay servers only.'
      operationId: storeWebhooksCinetpay
      requestBody:
        description: 'CinetPay webhook payload'
        required: true
        content:
          application/json:
            schema:
              properties:
                cpm_trans_id:
                  description: 'CinetPay transaction ID'
                  type: string
                  example: CP123456789
                cpm_custom:
                  description: 'Order number (custom field)'
                  type: string
                  example: ORD-123456
                cpm_amount:
                  description: 'Payment amount in FCFA'
                  type: number
                  format: float
                  example: 15000
                cpm_result:
                  description: "Payment result code ('00' = success)"
                  type: string
                  example: '00'
                signature:
                  description: 'HMAC signature = sha256(site_id + order_number + api_key)'
                  type: string
                  example: abc123def456...
                payment_method:
                  description: 'Payment provider used'
                  type: string
                  example: FLOOZ
              type: object
      responses:
        200:
          description: 'Webhook processed successfully'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                type: object
        400:
          description: 'Invalid signature'
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'Invalid signature'
                type: object
        404:
          description: 'Order not found'
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'Order not found'
                type: object
  '/api/webhooks/cinetpay/return/{order}':
    get:
      tags:
        - Webhooks
      summary: 'CinetPay payment return URL'
      description: 'User is redirected here after completing payment on CinetPay. Checks payment status and redirects to frontend success/failure page.'
      operationId: getWebhooksCinetpayReturn
      parameters:
        -
          name: order
          in: path
          description: 'Order UUID'
          required: true
          schema:
            type: string
            format: uuid
            example: 9d445a1c-85c5-4b6d-9c38-99a4915d6dac
      responses:
        302:
          description: 'Redirect to frontend success or failure page'
          headers:
            Location:
              description: 'Frontend URL'
              schema:
                type: string
                example: 'https://pourier.bf/orders/9d445a1c-85c5-4b6d-9c38-99a4915d6dac/success'
        404:
          $ref: '#/components/responses/NotFoundResponse'
components:
  schemas:
...
    ValidationErrorResponse:
      title: 'Validation Error Response'
      description: 'Réponse pour une erreur de validation'
      properties:
        message:
          type: string
          example: 'Les données fournies sont invalides'
        errors:
          description: 'Détails des erreurs de validation par champ'
          type: object
          example:
            email:
              - "L'email est invalide"
            password:
              - 'Le mot de passe est requis'
      type: object
    UnauthorizedResponse:
      title: 'Unauthorized Response'
      description: 'Réponse pour une authentification requise'
      properties:
        message:
          type: string
          example: 'Non authentifié'
      type: object
...
    NotFoundResponse:
      title: 'Not Found Response'
      description: 'Réponse pour une ressource non trouvée'
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: 'Ressource non trouvée'
      type: object
    PaginationMeta:
      title: 'Pagination Metadata'
      description: 'Métadonnées de pagination Laravel'
      properties:
        current_page:
          type: integer
          example: 1
        from:
          type: integer
          example: 1
        last_page:
          type: integer
          example: 10
        per_page:
          type: integer
          example: 15
        to:
          type: integer
          example: 15
        total:
          type: integer
          example: 150
      type: object
    PaginationLinks:
      title: 'Pagination Links'
      description: 'Liens de pagination Laravel'
      properties:
        first:
          type: string
          example: 'http://api.pourier.com/api/photos?page=1'
        last:
          type: string
          example: 'http://api.pourier.com/api/photos?page=10'
        prev:
          type: string
          example: null
          nullable: true
        next:
          type: string
          example: 'http://api.pourier.com/api/photos?page=2'
      type: object
...
    Category:
      title: Category
      description: 'Catégorie de photos'
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: Paysages
        slug:
          type: string
          example: paysages
        description:
          type: string
          example: 'Photos de paysages magnifiques'
          nullable: true
        icon:
          type: string
          example: landscape
          nullable: true
        photos_count:
          type: integer
          example: 1250
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
    Photo:
      title: Photo
      description: 'Modèle photo complet'
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        title:
          type: string
          example: 'Coucher de soleil sur la plage'
        description:
          type: string
          example: 'Un magnifique coucher de soleil capturé à Assinie'
          nullable: true
        slug:
          type: string
          example: coucher-de-soleil-sur-la-plage
        tags:
          type: array
          items:
            type: string
          example:
            - plage
            - 'coucher de soleil'
            - océan
            - nature
        preview_url:
          type: string
          example: 'https://storage.pourier.com/photos/preview.jpg'
        thumbnail_url:
          type: string
          example: 'https://storage.pourier.com/photos/thumbnail.jpg'
        watermarked_url:
          type: string
          example: 'https://storage.pourier.com/photos/watermarked.jpg'
        width:
          type: integer
          example: 4000
        height:
          type: integer
          example: 3000
        file_size:
          description: 'Taille en octets'
          type: integer
          example: 5242880
        format:
          type: string
          example: jpg
        exif_data:
          type: object
          example:
            camera: 'Canon EOS 5D'
            aperture: f/2.8
            iso: '100'
            shutter_speed: 1/500
          nullable: true
        color_palette:
          type: array
          items:
            type: string
          example:
            - '#FF6B35'
            - '#004E89'
            - '#1A659E'
        location:
          type: string
          example: "Assinie, Côte d'Ivoire"
          nullable: true
        price_standard:
          description: 'Prix en FCFA'
          type: number
          format: float
          example: 5000
        price_extended:
          description: 'Prix en FCFA'
          type: number
          format: float
          example: 15000
        status:
          type: string
          enum:
            - pending
            - approved
            - rejected
          example: approved
        views_count:
          type: integer
          example: 1250
        downloads_count:
          type: integer
          example: 45
        favorites_count:
          type: integer
          example: 89
        is_featured:
          type: boolean
          example: false
        category_id:
          type: string
          format: uuid
        photographer_id:
          type: string
          format: uuid
        photographer:
          $ref: '#/components/schemas/PhotographerProfile'
        category:
          $ref: '#/components/schemas/Category'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
    PhotographerProfile:
      title: 'Photographer Profile'
      description: "Profil public d'un photographe"
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Marie Kouassi'
        profile_picture:
          type: string
          nullable: true
        bio:
          type: string
          example: 'Photographe professionnelle basée à Abidjan'
          nullable: true
        photos_count:
          type: integer
          example: 125
        total_sales:
          type: integer
          example: 450
      type: object
    OrderItem:
      title: 'Order Item'
      description: "Article d'une commande"
      properties:
        id:
          type: string
          format: uuid
        photo_id:
          type: string
          format: uuid
        license_type:
          type: string
          enum:
            - standard
            - extended
          example: standard
        price:
          type: number
          format: float
          example: 5000
        photo:
          $ref: '#/components/schemas/Photo'
      type: object
    Order:
      title: Order
      description: 'Commande complète'
      properties:
        id:
          type: string
          format: uuid
        order_number:
          type: string
          example: ORD-20240115-001
        user_id:
          type: string
          format: uuid
        subtotal:
          type: number
          format: float
          example: 15000
        commission:
          description: '20% de commission'
          type: number
          format: float
          example: 3000
        total:
          type: number
          format: float
          example: 15000
        payment_status:
          type: string
          enum:
            - pending
            - completed
            - failed
            - refunded
          example: pending
        payment_method:
          type: string
          example: mobile_money
          nullable: true
        transaction_id:
          type: string
          example: CPAY-123456789
          nullable: true
        invoice_number:
          type: string
          example: INV-20240115-001
          nullable: true
        invoice_date:
          type: string
          format: date-time
          nullable: true
        invoice_path:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
  responses:
    UnauthorizedResponse:
      description: 'Unauthorized - Authentication required'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedResponse'
...
    NotFoundResponse:
      description: 'Resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFoundResponse'
    ValidationErrorResponse:
      description: 'Validation error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Enter JWT Bearer token obtained from /api/auth/login endpoint'
...
security:
  -
    bearerAuth: []